Remove AvailabilityView trait
